Add secrets-env to Nova MacOS job
