+++ .github/workflows/test_linux_job.yml
@@ -6,7 +6,7 @@ on:
       - .github/workflows/linux_job.yml
       - .github/workflows/test_linux_job.yml
       - .github/actions/setup-linux/action.yml
-      - .github/scripts/run_docker_with_env_secrets.py
+      - .github/scripts/run_with_env_secrets.py
   workflow_dispatch:
 
 jobs:
