+++ .github/workflows/test_macos_job.yml
@@ -5,6 +5,7 @@ on:
     paths:
       - .github/workflows/macos_job.yml
       - .github/workflows/test_macos_job.yml
+      - .github/scripts/run_with_env_secrets.py
   workflow_dispatch:
 
 jobs:
@@ -87,3 +88,24 @@ jobs:
       download-artifact: my-cool-artifact
       script: |
         grep  "hello" "${RUNNER_ARTIFACT_DIR}/cool_beans"
+  test-secrets-no-filter-var:
+    uses: ./.github/workflows/macos_job.yml
+    secrets: inherit
+    with:
+      job-name: "test-secrets-no-filter-var"
+      runner: macos-m1-12
+      test-infra-repository: ${{ github.repository }}
+      test-infra-ref: ${{ github.ref }}
+      script: |
+        [[ "${SECRET_NOT_A_SECRET_USED_FOR_TESTING}" == "SECRET_VALUE" ]] || exit 1
+  test-secrets-filter-var:
+    uses: ./.github/workflows/macos_job.yml
+    secrets: inherit
+    with:
+      job-name: "test-secrets-filter-var"
+      runner: linux.2xlarge
+      secrets-env: "NOT_A_SECRET_USED_FOR_TESTING"
+      test-infra-repository: ${{ github.repository }}
+      test-infra-ref: ${{ github.ref }}
+      script: |
+        [[ "${SECRET_NOT_A_SECRET_USED_FOR_TESTING}" == "SECRET_VALUE" ]] || exit 1
